backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.2.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Utilities
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
httpx>=0.26.0
aiosqlite>=0.19.0
